Positron Assistant: Reorganise participant commands
